*** empty log message ***
Tue Nov  5 00:04:48 1996  Thomas Bushnell, n/BSG  <thomas@gnu.ai.mit.edu>

	* stdio-common/vfprintf.c: Include <libc-lock.h>.

+++ sysdeps/mach/hurd/ttyname_r.c
@@ -1,4 +1,4 @@
-/* Copyright (C) 1994, 1995 Free Software Foundation, Inc.
+/* Copyright (C) 1994, 1995, 1996 Free Software Foundation, Inc.
 This file is part of the GNU C Library.
 
 The GNU C Library is free software; you can redistribute it and/or
@@ -26,7 +26,7 @@ Cambridge, MA 02139, USA.  */
 /* Store at most BUFLEN characters of the pathname of the terminal FD is
    open on in BUF.  Return 0 on success, -1 otherwise.  */
 int
-ttyname_r (int fd, char *buf, int buflen)
+ttyname_r (int fd, char *buf, size_t buflen)
 {
   error_t err;
   char nodename[1024];	/* XXX */
